feat: Content 화면 Navigation Controller 적용 및 Navigation Bar Setting
