+++ PageApp/PageApp/Base.lproj/Main.storyboard
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="19529" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="BYZ-38-t0r">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="19529" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="Cqg-Ue-wrK">
     <device id="retina6_1" orientation="portrait" appearance="light"/>
     <dependencies>
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="19519"/>
@@ -8,16 +8,16 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--Contents View Controller-->
+        <!--Content View Controller-->
         <scene sceneID="tne-QT-ifu">
             <objects>
-                <viewController id="BYZ-38-t0r" customClass="ContentsViewController" customModule="PageApp" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="BYZ-38-t0r" customClass="ContentViewController" customModule="PageApp" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="8bC-Xf-vdC">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
                             <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="wMU-rk-y2l">
-                                <rect key="frame" x="20" y="104" width="374" height="200"/>
+                                <rect key="frame" x="20" y="98" width="374" height="200"/>
                                 <subviews>
                                     <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="이벤트" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="jGZ-2Y-wiy">
                                         <rect key="frame" x="37" y="9" width="55" height="26"/>
@@ -57,7 +57,7 @@
                                 </constraints>
                             </view>
                             <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="28" estimatedSectionHeaderHeight="-1" sectionFooterHeight="28" estimatedSectionFooterHeight="-1" translatesAutoresizingMaskIntoConstraints="NO" id="zn3-aK-XSO">
-                                <rect key="frame" x="20" y="324" width="374" height="538"/>
+                                <rect key="frame" x="20" y="318" width="374" height="544"/>
                                 <prototypes>
                                     <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" preservesSuperviewLayoutMargins="YES" selectionStyle="default" indentationWidth="10" reuseIdentifier="BasicContentCell" rowHeight="104" id="JW4-gf-F23" customClass="ContentBasicCell" customModule="PageApp" customModuleProvider="target">
                                         <rect key="frame" x="0.0" y="44.5" width="374" height="104"/>
@@ -117,7 +117,7 @@
                         <viewLayoutGuide key="safeArea" id="6Tk-OE-BBY"/>
                         <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                         <constraints>
-                            <constraint firstItem="wMU-rk-y2l" firstAttribute="top" secondItem="6Tk-OE-BBY" secondAttribute="top" constant="60" id="FC1-KD-hjg"/>
+                            <constraint firstItem="wMU-rk-y2l" firstAttribute="top" secondItem="6Tk-OE-BBY" secondAttribute="top" constant="10" id="FC1-KD-hjg"/>
                             <constraint firstItem="zn3-aK-XSO" firstAttribute="leading" secondItem="wMU-rk-y2l" secondAttribute="leading" id="Kf6-5r-Zq6"/>
                             <constraint firstItem="zn3-aK-XSO" firstAttribute="bottom" secondItem="6Tk-OE-BBY" secondAttribute="bottom" id="Ub4-iw-gpt"/>
                             <constraint firstItem="zn3-aK-XSO" firstAttribute="trailing" secondItem="wMU-rk-y2l" secondAttribute="trailing" id="Uib-Sq-uKx"/>
@@ -126,11 +126,47 @@
                             <constraint firstItem="6Tk-OE-BBY" firstAttribute="trailing" secondItem="wMU-rk-y2l" secondAttribute="trailing" constant="20" id="l5C-DJ-gNz"/>
                         </constraints>
                     </view>
+                    <navigationItem key="navigationItem" id="zBp-cn-wnX"/>
                     <connections>
                         <outlet property="contentTableView" destination="zn3-aK-XSO" id="Dy3-Yl-vdL"/>
+                        <segue destination="f54-sH-Kas" kind="show" identifier="moveCartVC" id="cdU-Ap-PAr"/>
                     </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="dkx-z0-nzr" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="928.98550724637687" y="97.767857142857139"/>
+        </scene>
+        <!--View Controller-->
+        <scene sceneID="Ms7-wP-fHS">
+            <objects>
+                <viewController id="f54-sH-Kas" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="WrS-fJ-c2k">
+                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <viewLayoutGuide key="safeArea" id="rAD-iM-zUN"/>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                    </view>
+                    <navigationItem key="navigationItem" id="ZxG-Ul-7kF"/>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="YXz-nV-c4w" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="1871" y="98"/>
+        </scene>
+        <!--Navigation Controller-->
+        <scene sceneID="Qvu-y3-TL3">
+            <objects>
+                <navigationController automaticallyAdjustsScrollViewInsets="NO" id="Cqg-Ue-wrK" sceneMemberID="viewController">
+                    <toolbarItems/>
+                    <navigationBar key="navigationBar" contentMode="scaleToFill" id="E3P-Qi-bUa">
+                        <rect key="frame" x="0.0" y="44" width="414" height="44"/>
+                        <autoresizingMask key="autoresizingMask"/>
+                    </navigationBar>
+                    <nil name="viewControllers"/>
+                    <connections>
+                        <segue destination="BYZ-38-t0r" kind="relationship" relationship="rootViewController" id="5Pj-km-X40"/>
+                    </connections>
+                </navigationController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="nug-vs-VTH" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
             <point key="canvasLocation" x="18.840579710144929" y="97.767857142857139"/>
         </scene>
