--- PageApp/PageApp/content/ContentsViewController.swift
@@ -1,50 +1,0 @@
-//
-//  ViewController.swift
-//  PageApp
-//
-//  Created by river.o on 2022/01/25.
-//
-
-import UIKit
-
-class ContentsViewController: UIViewController, UITableViewDelegate {
-    
-    @IBOutlet weak var contentTableView: UITableView!
-    
-    private let contentDataSource = ContentDataSource()
-    lazy var contents = Content.jsonToContents()
-        
-    override func viewDidLoad() {
-        super.viewDidLoad()
-       
-        initView()
-        initTableView()
-    }
-    
-    private func initView() {
-        initHeader()
-    }
-    
-    private func initHeader() {
-        let headerView: HeaderView = {
-            let view = HeaderView()
-            view.translatesAutoresizingMaskIntoConstraints = false
-            return view
-        }()
-        
-        view.addSubview(headerView)
-
-        headerView.leadingAnchor.constraint(equalTo: view.leadingAnchor).isActive = true
-        headerView.trailingAnchor.constraint(equalTo: view.trailingAnchor).isActive = true
-        headerView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor).isActive = true
-        headerView.widthAnchor.constraint(equalToConstant: view.frame.width).isActive = true
-        headerView.heightAnchor.constraint(equalToConstant: 50).isActive = true
-    }
-    
-    private func initTableView() {
-        contentTableView.delegate = self
-        contentTableView.dataSource = contentDataSource
-        contentDataSource.contents = contents
-    }
-}
-
